A Tweak of Case-insens Lookup in IDispatchExImpl
When adding a series of variable names with the same letters but in different
cases, The AddEntryToMaps() overwrites the old one with the new one in the case
Insensitive search map mNameToEntryIMap.

This patch changes the behavior to the one described below.

the MSFT implementation of IDispatchEx always preserves the first variable
name added to the case insensitive search map. Subsequence additions are
ignored.